cleaned up code for Space Scene
