+++ public/index.html
@@ -6,8 +6,6 @@
   <meta charset="UTF-8">
   <link rel="stylesheet" href="styles/normalize.css"/> 
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootswatch/3.3.7/cyborg/bootstrap.min.css"/>
-  <!-- idk why this link isn't working -->
-  <!-- <link rel="stylesheet" href="../../node_modules/bootstrap/bootstrap.min.css"/> -->
   <link rel="stylesheet" href="styles/style.css"/>
   <script src="js/three.js"></script>
   <script src="https://aframe.io/releases/0.5.0/aframe.min.js"></script>
